requirement.txt
asttokens==2.4.1
backcall==0.2.0
cffi==1.17.0
colorama==0.4.6
comm==0.2.2
cryptography==43.0.0
debugpy==1.8.6
decorator==5.1.1
executing==2.1.0
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.1
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
pickleshare==0.7.5
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
pure_eval==0.2.3
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
pywin32==306
pyzmq==26.2.0
six==1.16.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
zipp==3.20.2
